requirements.txt
selenium==4.22.0
beautifulsoup4==4.12.3
requests==2.32.3
SQLAlchemy==2.0.31
pandas==2.2.2
matplotlib==3.9.1
tldextract==5.1.2
python-dotenv==1.0.1
webdriver-manager==4.0.1